requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
openai>=1.0.0
Flask-Cors